SI-4492 feat(notes): Файл с инструментами для работы с темами
