+++ draggable_notes/lib/res/themes.dart
@@ -44,16 +44,3 @@ class AppThemes {
     ),
   );
 }
-
-extension StringTheme on String {
-  /// Получить режим темы по строке
-  ThemeMode toThemeMode() {
-    if (this == 'light') {
-      return ThemeMode.light;
-    }
-    if (this == 'dark') {
-      return ThemeMode.dark;
-    }
-    return ThemeMode.system;
-  }
-}
