+++ draggable_notes/lib/ui/notes/notes_screen.dart
@@ -1,5 +1,4 @@
 import 'package:draggable_notes/data/note.dart';
-import 'package:draggable_notes/res/colors.dart';
 import 'package:draggable_notes/res/strings.dart';
 import 'package:draggable_notes/ui/notes/notes_wm.dart';
 import 'package:draggable_notes/ui/widgets/create_task_button.dart';
