+++ draggable_notes/lib/ui/widgets/create_task_button.dart
@@ -1,4 +1,3 @@
-import 'package:draggable_notes/res/colors.dart';
 import 'package:draggable_notes/res/strings.dart';
 import 'package:flutter/material.dart';
