+++ draggable_notes/lib/ui/widgets/note_card.dart
@@ -1,4 +1,3 @@
-import 'package:draggable_notes/res/colors.dart';
 import 'package:flutter/material.dart';
 
 /// Карточка с заметкой
