+++ draggable_notes/pubspec.yaml
@@ -13,15 +13,15 @@ dependencies:
   elementary: ^2.0.0
   flutter:
     sdk: flutter
+  get_it: ^7.6.0
+  golden_toolkit: ^0.15.0
   hive_flutter: ^1.1.0
   hive_generator: ^2.0.0
-  surf_lint_rules: ^3.0.0
-  get_it: ^7.6.0
   injectable: ^2.1.2
-  golden_toolkit: ^0.15.0
   mocktail: ^0.3.0
   provider: ^6.0.5
   shared_preferences: ^2.2.0
+  surf_lint_rules: ^3.0.0
 
 dev_dependencies:
   build_runner: ^2.4.6
