+++ draggable_notes/test/golden_test.dart
@@ -34,7 +34,7 @@ void main() {
 
     await tester.pumpWidgetBuilder(
       builder.build(),
-      surfaceSize: Size(400, 800),
+      surfaceSize: const Size(400, 800),
     );
 
     await screenMatchesGolden(tester, 'notes_card');
@@ -44,7 +44,7 @@ void main() {
     final builder = GoldenBuilder.column()
       ..addScenario(
         'create_new_note_dialog',
-        CreateNoteDialog(),
+        const CreateNoteDialog(),
       );
 
     await tester.pumpWidgetBuilder(builder.build());
